chore: isolate legacy route resolver
